+++ CHANGELOG.md
@@ -14,9 +14,15 @@ and `Removed`.
 
 ## [Unreleased]
 
-## Added
+### Added
 
+- Ability to select compression level for when using Zstd backups.
+- Ability to backup Fonts folder.
 - SavedVariables can be removed for a single addon under details.
+
+### Fixed
+
+- Missing translations
 
 ## [1.2.2] - 2021-07-06